Fix inconsistent language in some cases
In these situations, the language from the user was used instead of the language
from the request. That meant, if the cookie and the value on the server did not
match, some parts (here the problem-titles) were in another language than the
rest of the site. This especially happened when working on staging and
production simultaneous, since they are using the same cookie.